requirements.txt
PyMuPDF
langchain
sentence-transformers
faiss-cpu
numpy